Add files via upload
Current test bench C code for zero crossing. I have totally changed how the initial function to populate the struct will work to help with the zero cross detector. Mostly to keep track of number of peaks and mins for the sampled signal, to then help with the conditions used in the zero detector function. I am using an arbitrary array that mimics a 1.5 cycle sine wave with 2 peaks 1 min, or 2 mins and 1 peak for checking the logic works correctly.